% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodings.R
\name{encode_deviation}
\alias{encode_deviation}
\title{Encode a given factor variable using deviation encoding}
\usage{
encode_deviation(X, fact, keep_factor = FALSE)
}
\arguments{
\item{X}{The data.frame/data.table to transform.}

\item{fact}{The factor variable to encode by - either a positive integer specifying the 
column number, or the name of the column.}

\item{keep_factor}{Whether to keep the original factor column(defaults to **FALSE**).}
}
\value{
A new data.table X which contains the new columns and optionally the old factor.
}
\description{
Transforms the original design matrix using a deviation dummy encoding.
}
\details{
The deviation dummy variable encoding, with reference class level set to -1. 
The reference class is always the last class observed.
}
\examples{

design_mat <- cbind( data.frame( matrix(rnorm(5*100),ncol = 5) ),
                     sample( sample(letters, 10), 100, replace = TRUE)
                     )
colnames(design_mat)[6] <- "factor_var"

#encode_deviation(X = design_mat, fact = "factor_var", keep_factor = FALSE)


}
