% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodings.R
\name{encode_mnl}
\alias{encode_mnl}
\title{Encode a given factor variable using a multinomial logit representation}
\usage{
encode_mnl(X, fact, keep_factor = FALSE)
}
\arguments{
\item{X}{The data.frame/data.table to transform.}

\item{fact}{The factor variable to encode by - either a positive integer specifying the 
column number, or the name of the column.}

\item{keep_factor}{Whether to keep the original factor column(defaults to **FALSE**).}
}
\value{
A new data.table X which contains the new columns and optionally the old factor.
}
\description{
Transforms the original design matrix using a mnl encoding.
}
\details{
Uses the method from Johannemann et al.(2019) 
'Sufficient Representations for Categorical Variables' - mnl.
}
\examples{

design_mat <- cbind( data.frame( matrix(rnorm(5*100),ncol = 5) ),
                     sample( sample(letters, 10), 100, replace = TRUE)
                     )
colnames(design_mat)[6] <- "factor_var"

encode_mnl(X = design_mat, fact = "factor_var", keep_factor = FALSE)


}
